refactor: nwc module
