+++ fedimint-nwc/src/nwc/mod.rs
@@ -1,0 +1,15 @@
+use nostr::nips::nip47::Method;
+
+pub mod handlers;
+pub mod profiles;
+
+pub const METHODS: [Method; 8] = [
+    Method::GetInfo,
+    Method::MakeInvoice,
+    Method::GetBalance,
+    Method::LookupInvoice,
+    Method::PayInvoice,
+    Method::MultiPayInvoice,
+    Method::PayKeysend,
+    Method::MultiPayKeysend,
+];
