+++ fedimint-nwc/src/nwc/profiles.rs
@@ -1,0 +1,1 @@
+
